+++ .pre-commit-config.yaml
@@ -49,11 +49,11 @@ repos:
         language: system
         pass_filenames: false
         always_run: true
-  - repo: local
-    hooks:
-      - id: trufflehog
-        name: TruffleHog
-        description: Detect secrets in your data.
-        entry: bash -c 'trufflehog git file://. --no-verification --fail --exclude-paths=.trufflehogignore'
-        language: system
-        stages: ["commit", "push"]
+  # - repo: local
+  #   hooks:
+  #     - id: trufflehog
+  #       name: TruffleHog
+  #       description: Detect secrets in your data.
+  #       entry: bash -c 'trufflehog git file://. --no-verification --fail --exclude-paths=.trufflehogignore'
+  #       language: system
+  #       stages: ["commit", "push"]
